update redisgn animal so dogs and cats are separated
